api/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pillow==11.0.0
pytest==8.3.3
python-multipart==0.0.12

httpx==0.27.2
